+++ src/CronBuilderServiceProvider.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace InterwalNet\CronBuilder;
 
+use Filament\Support\Assets\Css;
+use Filament\Support\Facades\FilamentAsset;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -17,5 +19,12 @@ class CronBuilderServiceProvider extends PackageServiceProvider
             ->hasConfigFile()
             ->hasTranslations()
             ->hasViews();
+    }
+
+    public function packageBooted(): void
+    {
+        FilamentAsset::register([
+            Css::make('cron-builder', __DIR__.'/../resources/css/cron-builder.css'),
+        ], package: 'interwal-net/filament-visual-cron-builder');
     }
 }